src/da/dev_requirements.txt
#releasing
urllib.request
json
ssl
datetime
mysql.connector
pytz
netCDF4
aqi
os
cdsapi
csv